Move a trait and singleton around.
